Extract command factory creation
