+++ server/src/main/java/io/spine/system/server/DefaultSystemGateway.java
@@ -22,14 +22,10 @@ package io.spine.system.server;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Message;
-import io.spine.client.ActorRequestFactory;
 import io.spine.client.CommandFactory;
 import io.spine.core.Command;
-import io.spine.core.TenantId;
 import io.spine.core.UserId;
 import io.spine.server.BoundedContext;
-import io.spine.server.tenant.TenantFunction;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.grpc.StreamObservers.noOpObserver;
@@ -44,9 +40,9 @@ import static io.spine.grpc.StreamObservers.noOpObserver;
 final class DefaultSystemGateway implements SystemGateway {
 
     /**
-     * The which posts the system events.
+     * The ID of the user used for generating system commands and events.
      */
-    private static final UserId SYSTEM = UserId
+    static final UserId SYSTEM_USER = UserId
             .newBuilder()
             .setValue("SYSTEM")
             .build();
@@ -60,7 +56,8 @@ final class DefaultSystemGateway implements SystemGateway {
     @Override
     public void postCommand(Message systemCommand) {
         checkNotNull(systemCommand);
-        CommandFactory commandFactory = buildRequestFactory().command();
+        CommandFactory commandFactory =
+                TenantAwareCommandFactory.newInstance(system.isMultitenant());
         Command command = commandFactory.create(systemCommand);
         system.getCommandBus()
               .post(command, noOpObserver());
@@ -74,36 +71,5 @@ final class DefaultSystemGateway implements SystemGateway {
     @VisibleForTesting
     BoundedContext target() {
         return system;
-    }
-
-    private ActorRequestFactory buildRequestFactory() {
-        return system.isMultitenant()
-               ? buildMultitenantFactory()
-               : buildSingleTenantFactory();
-    }
-
-    private static ActorRequestFactory buildMultitenantFactory() {
-        TenantFunction<ActorRequestFactory> contextFactory =
-                new TenantFunction<ActorRequestFactory>(true) {
-                    @Override
-                    public ActorRequestFactory apply(@Nullable TenantId tenantId) {
-                        checkNotNull(tenantId);
-                        return constructFactory(tenantId);
-                    }
-                };
-        ActorRequestFactory result = contextFactory.execute();
-        checkNotNull(result);
-        return result;
-    }
-
-    private static ActorRequestFactory buildSingleTenantFactory() {
-        return constructFactory(TenantId.getDefaultInstance());
-    }
-
-    private static ActorRequestFactory constructFactory(TenantId tenantId) {
-        return ActorRequestFactory.newBuilder()
-                                  .setActor(SYSTEM)
-                                  .setTenantId(tenantId)
-                                  .build();
     }
 }
